Move VertexArray to new file
