+++ src/client/CMakeLists.txt
@@ -1,7 +1,7 @@
 
 add_library(ob-client
     input/keyboard.cpp
-    gl/gl_vertex_array.cpp
+    gl/vertex_array.cpp
     gl/texture_atlas.cpp
     gl/gl_shader.cpp
     gl/gl_textures.cpp
